Moved demo data generator to new file.
