+++ tools/graphing/demo_data_gen.py
@@ -1,0 +1,45 @@
+import numpy.random as random
+
+
+def get_demo_data():
+    """
+    Generates random data in the format the grapher will receive data.
+
+    Returns
+    -------
+    dict
+        Keys of data streams w/ randomly generated data.
+    """
+
+    gen = random.rand(26, 1)
+
+    imaginary_data = {
+        'altitude': gen[0, 0],
+        'airspeed': gen[1, 0],
+        'velocity_x': gen[2, 0],
+        'velocity_y': gen[3, 0],
+        'velocity_z': gen[4, 0],
+        'voltage': gen[5, 0],
+        'state': gen[6, 0],
+        'mode': gen[7, 0],
+        'armed': gen[8, 0],
+        'roll': gen[9, 0],
+        'pitch': gen[10, 0],
+        'yaw': gen[11, 0],
+        'altitude_controller_output': gen[12, 0],
+        'altitude_rc_output': gen[13, 0],
+        'target_altitude': gen[14, 0],
+        'pitch_controller_output': gen[15, 0],
+        'pitch_rc_output': gen[16, 0],
+        'target_pitch_velocity': gen[17, 0],
+        'roll_controller_output': gen[18, 0],
+        'roll_rc_output': gen[19, 0],
+        'target_roll_velocity': gen[20, 0],
+        'yaw_controller_output': gen[21, 0],
+        'yaw_rc_output': gen[22, 0],
+        'target_yaw': gen[23, 0],
+        'color_image': gen[24, 0],
+        'depth_image': gen[25, 0]
+    }
+
+    return imaginary_data
